requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
pymongo==3.11.3
PyJWT==2.0.1
dnspython==2.0.0
gunicorn==19.9.0